refactor: remove unnecessary imports and better organize
